Handle releases with multiple zips
The previous code would simple use the "last" one that Github returned, but this may not be the coreect one.  Instead use the largest one, which should always be the OTA zip file.

This is useful for things like including other zip files in a release like the recovery image, or vbmeta, or pkmd files.